refactor: Password 로직 인터페이스 화 삭제
